Add: JWT Filter, and use @AuthenticationPrincipal, serperate  Auth from User

+++ src/main/java/com/dna/fooo_guard/domain/user/UserService.java
@@ -3,9 +3,9 @@ package com.dna.fooo_guard.domain.user;
 import java.util.List;
 
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
 
 import com.dna.fooo_guard.domain.user.dto.UserResponse;
-import com.dna.fooo_guard.domain.user.dto.UserSignUpRequest;
 import com.dna.fooo_guard.global.error.CustomException;
 import com.dna.fooo_guard.global.error.ErrorCode;
 
@@ -13,38 +13,34 @@ import lombok.RequiredArgsConstructor;
 
 @Service
 @RequiredArgsConstructor
+@Transactional
 public class UserService {
     private final UserRepository userRepository;
-
-    // TODO: request말고 dto로 바꿀지 고민
-    public String signUp(UserSignUpRequest request) {
-        if (userRepository.existsByNickname(request.getNickname())) {
-            throw new CustomException(ErrorCode.DUPLICATE_NICKNAME);
-        }
-
-        User user = User.builder()
-                .username(request.getUsername())
-                .password(request.getPassword())
-                .nickname(request.getNickname())
-                .build();
-        userRepository.save(user);
-        return String.format("유저[%s] - 회원가입", user.getUsername());
-    }
-
+    
+    @Transactional(readOnly=true)
     public UserResponse findUserById(Long id) {
         User user = userRepository.findById(id)
-                .orElseThrow(() -> new IllegalArgumentException("User not found"));
+                .orElseThrow(() -> new CustomException(ErrorCode.USER_NOT_FOUND));
         return UserResponse.from(user);
     }
 
+    @Transactional(readOnly=true)
     public List<UserResponse> findAllUsersByGroupId(Long groupId) {
         List<User> users = userRepository.findAllByGroupId(groupId);
+
+        if (users.isEmpty()) {
+            throw new CustomException(ErrorCode.USER_NOT_FOUND);
+        }
+
         return users.stream()
                 .map(UserResponse::from)
                 .toList();
     }
 
     public void deleteUser(Long id) {
+        if (!userRepository.existsById(id)) {
+            throw new CustomException(ErrorCode.USER_NOT_FOUND);
+        }
         userRepository.deleteById(id);
     }
 }
